Add new arguments for cross compiling
A toolchain file can be used to specify compiler, flags and system for
cross compiling. Since the include of the ForceCompiler CMake module
did not work, we used the commands specified in the module directly.
Among other things, this disables the cmake compiler check.

The CMake file implements an additional compiler check to check for c++11,
which does not work with the arm-none-eabi cross compiler, a new flag
was added to skip this check.